requirements.txt
grpcio
grpcio-tools
pyzmq
mysql
mysql-connector-python